django_project/pyproject.toml
[tool.poetry]
name = "geoquizz_backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2"
psycopg2 = "^2.8.6"
djangorestframework = "^3.12.4"
drf-yasg = "^1.20.0"
Pillow = "^8.2.0"
factory-boy = "^3.2.0"
geojson = "^2.5.0"
djangorestframework-gis = "^0.17"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
